+++ src/echelon3/metrics/detection.py
@@ -24,9 +24,12 @@ class mAP(Metric):
 
     engine = None
 
-    def __init__(self, bbox_type='yolo'):
+    def __init__(self, bbox_type='yolo', backend='faster_coco_eval'):
         self.bbox_type = BBoxType(bbox_type)
-        self.engine = mean_ap.MeanAveragePrecision(box_format=BBOX_TYPE_TO_FORMAT[self.bbox_type])
+        self.engine = mean_ap.MeanAveragePrecision(
+            box_format=BBOX_TYPE_TO_FORMAT[self.bbox_type],
+            backend=backend,
+        )
         self.reset()
 
     def update(self, predicted: torch.Tensor, target: torch.Tensor):
@@ -34,11 +37,11 @@ class mAP(Metric):
         device = predicted['heatmaps'].device
         addon_predicted_list = [{'boxes': coords_to_absolute(predicted['bboxes'][idx, :, :], output_size).to(device),
                                  'scores': torch.squeeze(predicted['scores'][idx, :, :], dim=-1).to(device),
-                                 'labels': torch.squeeze(predicted['classes'][idx, :, :], dim=-1).to(device)}
+                                 'labels': torch.squeeze(predicted['classes'][idx, :, :], dim=-1).long().to(device)}
                                 for idx in range(len(target))]
 
         addon_target_list = [{'boxes': coords_to_absolute(torch.Tensor([l[0:4] for l in target[idx]]), output_size).to(device),
-                              'labels': torch.Tensor([l[4] for l in target[idx]]).to(device)} for idx in range(len(target))]
+                              'labels': torch.tensor([int(l[4]) for l in target[idx]], dtype=torch.long).to(device)} for idx in range(len(target))]
 
         self.engine.update(addon_predicted_list, addon_target_list)